# refine softmax kernel