refine sofmtax:

bs1, num_class1000， 2.7147 => 1.2598us,  2.15X
bs10, num_class1000,  26.9311 => 12.0637us, 2.23X
on v3.

```
bs:1
...
I0128 18:00:37.661339 21036 benchmark.cc:156] Kernel Type kSoftmax: 995: Refer takes 15.8576 us; MKL takes 2.7021 us; Mixed takes 1.269 us; Target takes 2.7033 us; 
I0128 18:00:37.889328 21036 benchmark.cc:156] Kernel Type kSoftmax: 996: Refer takes 15.8707 us; MKL takes 2.7067 us; Mixed takes 1.2601 us; Target takes 2.7146 us; 
I0128 18:00:38.117602 21036 benchmark.cc:156] Kernel Type kSoftmax: 997: Refer takes 15.8868 us; MKL takes 2.7179 us; Mixed takes 1.2636 us; Target takes 2.7095 us; 
I0128 18:00:38.346132 21036 benchmark.cc:156] Kernel Type kSoftmax: 998: Refer takes 15.917 us; MKL takes 2.7118 us; Mixed takes 1.26362 us; Target takes 2.7125 us; 
I0128 18:00:38.574906 21036 benchmark.cc:156] Kernel Type kSoftmax: 999: Refer takes 15.9287 us; MKL takes 2.71632 us; Mixed takes 1.2662 us; Target takes 2.71753 us; 
I0128 18:00:38.803541 21036 benchmark.cc:156] Kernel Type kSoftmax: 1000: Refer takes 15.9284 us; MKL takes 2.7147 us; Mixed takes 1.2598 us; Target takes 2.7134 us; 

bs:10
...
I0128 18:56:39.800736  4208 benchmark.cc:156] Kernel Type kSoftmax: 997: Refer takes 159.055 us; MKL takes 27.0435 us; Mixed takes 13.2519 us; Target takes 26.9862 us; 
I0128 18:56:42.087453  4208 benchmark.cc:156] Kernel Type kSoftmax: 998: Refer takes 159.157 us; MKL takes 27.0799 us; Mixed takes 13.0731 us; Target takes 27.0652 us; 
I0128 18:56:44.377470  4208 benchmark.cc:156] Kernel Type kSoftmax: 999: Refer takes 159.312 us; MKL takes 27.0701 us; Mixed takes 13.2842 us; Target takes 27.0394 us; 
I0128 18:56:46.650840  4208 benchmark.cc:156] Kernel Type kSoftmax: 1000: Refer takes 159.175 us; MKL takes 26.9311 us; Mixed takes 12.0637 us; Target takes 26.8871 us; 
```

how to run:
`./paddle/fluid/operators/jit/jit_kernel_benchmark  --burning=100 --repeat=10000 --max_size=1000 --filter=kSoftmax`