@tensor-tang 2019-01-29
bert 大数据集 提升26%
`thread0::softmax            130         54851       0.025667    733.663     421.931     0.156064`
=>
`thread0::softmax            130         43383.6     0.024485    375.366     333.72      0.106399`